+++ src/main/java/com/example_projec/SpringbootSwegger/controllers/ControllerMahasiswa.java
@@ -2,7 +2,11 @@ package com.example_projec.SpringbootSwegger.controllers;
 
 import com.example_projec.SpringbootSwegger.models.Mahasiswa;
 import com.example_projec.SpringbootSwegger.services.MahasiswaService;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
@@ -23,18 +27,25 @@ public class ControllerMahasiswa {
     }
 
     @PostMapping
-    public void addMahasiswa(@RequestBody Mahasiswa mahasiswa) {
-       mahasiswaService.addMahasiswa(mahasiswa);
+    public Mahasiswa addMahasiswa(@RequestBody Mahasiswa mahasiswa) {
+        mahasiswaService.addMahasiswa(mahasiswa);
+        return mahasiswa;
     }
 
-    @PutMapping("/{nim}")
-    public void updateMahasiswa(@PathVariable Long nim, @RequestBody Mahasiswa mahasiswa) {
-        mahasiswaService.updateMahasiswa(nim, mahasiswa);
+    @PutMapping("")
+    public Mahasiswa updateMahasiswa(@RequestBody Mahasiswa mahasiswa) {
+        mahasiswaService.updateMahasiswa(mahasiswa);
+        return mahasiswa;
     }
 
     @DeleteMapping("/{nim}")
-    public void deleteMahasiswa(@PathVariable Long nim) {
-        mahasiswaService.deleteMahasiswa(nim);
+    public Mahasiswa deleteMahasiswa(@PathVariable Long nim) {
+        Mahasiswa dt = mahasiswaService.getByNim(nim);
+        if(dt != null){
+            mahasiswaService.deleteMahasiswa(nim);
+        }
+        return dt;
     }
+
 }
